+++ pkgs/native_assets_builder/test/build_runner/concurrency_test_helper.dart
@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:file/local.dart';
 import 'package:logging/logging.dart';
 import 'package:native_assets_builder/native_assets_builder.dart';
 
@@ -25,6 +26,7 @@ void main(List<String> args) async {
     logger: logger,
     dartExecutable: dartExecutable,
     singleHookTimeout: timeout,
+    fileSystem: const LocalFileSystem(),
   ).build(
     configCreator: () => BuildConfigBuilder()
       ..setupCodeConfig(